Began refactoring to merge newcfg script into autocrab program.
