+++ autocrab/commands/status.py
@@ -1,0 +1,1 @@
+# I'll see if I can get this command working first.
