+++ autocrab/crabutil.py
@@ -1,0 +1,45 @@
+"""
+Utility functions for working with CRAB. Used by most
+of the autocrab commands.
+"""
+
+def getCrabConfig(dir):
+	"""Retrieves backup copy of crab config file in directory."""
+	path = os.path.join(dir, "share", "crab.cfg")
+	return path
+
+def getCrabLog(dir):
+	path = os.path.join(dir, "log", "crab.log")
+	return path
+
+def isCrabDirectory(dir):
+	"""Check if a directory is a working directory generated by crab."""
+	isDirectory = True
+	if not os.path.exists(os.path.join(dir, "job")):
+		isDirectory = False
+	if not os.path.exists(os.path.join(dir, "res")):
+		isDirectory = False
+	if not os.path.exists(os.path.join(dir, "log")):
+		isDirectory = False
+	if not os.path.exists(os.path.join(dir, "share")):
+		isDirectory = False
+	return isDirectory
+
+def isCrabConfig(filename):
+	"""Check if a file is a legal crab config file."""
+	valid = False
+	if ".cfg" in filename:
+		valid = True
+
+	configFile = open(filename)
+	text = configFile.read()
+
+	# Make sure we have the right blocks inside.
+	if not ("[CRAB]" in text and "[CMSSW]" in text and "[USER]" in text):
+		valid = False
+
+	# Make sure we're not an autocrab template config file
+	if "$INS_SHORTNAME" in text or "$INS_LONGNAME" in text:
+		valid = False	
+
+	return valid
